LAB02/requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
gunicorn==22.0.0
pydantic==2.8.2
pydantic-settings==2.4.0
PyYAML==6.0.2
python-dotenv==1.0.1
pandas==2.2.2
